Adds server action tests, makes fieldName configurable (#65)
