+++ packages/express/README.md
@@ -150,6 +150,7 @@ app.listen(port, () => {
   saltByteLength: 8,
   secretByteLength: 18,
   token: {
+    fieldName: 'csrf_token',
     responseHeader: 'X-CSRF-Token'
   }
 }
