+++ packages/nextjs/src/index.ts
@@ -2,8 +2,9 @@ import type { NextRequest } from 'next/server';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { NextResponse } from 'next/server';
 
-import { CsrfError, createCsrfProtect as _createCsrfProtect, Config, TokenOptions } from '@shared/protect';
-import type { ConfigOptions } from '@shared/protect';
+import { Config, TokenOptions } from '@shared/config';
+import type { ConfigOptions } from '@shared/config';
+import { CsrfError, createCsrfProtect as _createCsrfProtect } from '@shared/protect';
 
 export { CsrfError };
