+++ packages/sveltekit/src/index.ts
@@ -1,7 +1,8 @@
 import type { Handle, RequestEvent, Cookies } from '@sveltejs/kit';
 
-import { CsrfError, createCsrfProtect as _createCsrfProtect, Config, TokenOptions } from '@shared/protect';
-import type { ConfigOptions } from '@shared/protect';
+import { Config, TokenOptions } from '@shared/config';
+import type { ConfigOptions } from '@shared/config';
+import { CsrfError, createCsrfProtect as _createCsrfProtect } from '@shared/protect';
 
 export { CsrfError };
