+++ shared/src/protect.ts
@@ -1,87 +1,11 @@
+import { Config } from './config';
+import type { ConfigOptions, CookieOptions } from './config';
 import { createSecret, createToken, getTokenString, verifyToken, atou, utoa } from './util';
-import type { TokenValueFunction } from './util';
 
 /**
  * Represents a generic CSRF protection error
  */
 export class CsrfError extends Error {}
-
-/**
- * Represents cookie options in config
- */
-export class CookieOptions {
-  domain: string = '';
-
-  httpOnly: boolean = true;
-
-  maxAge: number | undefined = undefined;
-
-  name: string = '_csrfSecret';
-
-  partitioned: boolean | undefined = undefined;
-
-  path: string = '/';
-
-  sameSite: boolean | 'none' | 'strict' | 'lax' = 'strict';
-
-  secure: boolean = true;
-
-  constructor(opts?: Partial<CookieOptions>) {
-    Object.assign(this, opts);
-  }
-}
-
-/**
- * Represents token options in config
- */
-export class TokenOptions {
-  value: TokenValueFunction | undefined = undefined;
-
-  constructor(opts?: Partial<TokenOptions>) {
-    Object.assign(this, opts);
-  }
-}
-
-/**
- * Represents CsrfProtect configuration object
- */
-export class Config {
-  excludePathPrefixes: string[] = [];
-
-  ignoreMethods: string[] = ['GET', 'HEAD', 'OPTIONS'];
-
-  saltByteLength: number = 8;
-
-  secretByteLength: number = 18;
-
-  cookie: CookieOptions = new CookieOptions();
-
-  token: TokenOptions = new TokenOptions();
-
-  constructor(opts?: Partial<ConfigOptions>) {
-    const newOpts = opts || {};
-    if (newOpts.cookie) newOpts.cookie = new CookieOptions(newOpts.cookie);
-    if (newOpts.token) newOpts.token = new TokenOptions(newOpts.token);
-    Object.assign(this, newOpts);
-
-    // basic validation
-    if (this.saltByteLength < 1 || this.saltByteLength > 255) {
-      throw Error('saltBytLength must be greater than 0 and less than 256');
-    }
-
-    if (this.secretByteLength < 1 || this.secretByteLength > 255) {
-      throw Error('secretBytLength must be greater than 0 and less than 256');
-    }
-  }
-}
-
-/**
- * Represents CsrfProtect configuration options object
- */
-export interface ConfigOptions extends Omit<Config, 'cookie' | 'token'> {
-  cookie: Partial<CookieOptions>;
-  token: Partial<TokenOptions>;
-}
 
 /**
  * Represents a cookie
@@ -137,7 +61,7 @@ export function createCsrfProtect(opts?: Partial<ConfigOptions>): CsrfProtect {
 
     // verify token
     if (!config.ignoreMethods.includes(request.method)) {
-      const tokenStr = await getTokenString(request, config.token.value);
+      const tokenStr = await getTokenString(request, config.token);
       if (!await verifyToken(atou(tokenStr), secret)) {
         throw new CsrfError('csrf validation error');
       }
